PythonProject/requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
telethon==1.39.0
psycopg2==2.9.10
sqlalchemy==2.0.40
python-dotenv==1.1.0